Improved thread management
Due to the fact that dbping is a tool for generating concurrency, it is not interesting to use a threadPool to limit the number of tasks executed concurrently. For this reason, the threadPool has been replaced by direct invocations of the Thread class to raise the number of threads indicated in the configuration XML file.