New feature: repeatable commands
It's now possible to set each command to repeat a number of times before moving on to the execution of the next command. This functionality is especially useful, for example, to fill a temporary table to be used by a subsequent command. The execution of these command repetitions can be done one at a time or in blocks of a specific size.

To adapt the behavior of this new functionality, the "query" element allows you to configure the "repeats" and "batchSize" properties as follows:

<query label="insert_gtt" repeats="1000" batchSize="-1">
    <sql><![CDATA[INSERT INTO gtt_shop VALUES (?)]]></sql>
    <parameter order="1" type="INTEGER" file="/path/to/shops.csv"/>
</query>